Remove migration driver
- Remove unused migration driver
- Rename test exclusive schema files and functions to more clearly
  indicate their purpose
- Rename schema migration functions that only migrate local builtins
  rather than those from the module-index
