+++ lib/dal/src/lib.rs
@@ -242,7 +242,7 @@ pub async fn migrate(pg: &PgPool) -> ModelResult<()> {
 
 #[allow(clippy::too_many_arguments)]
 #[instrument(skip_all)]
-pub async fn migrate_builtins(
+pub async fn migrate_local_builtins(
     pg: &PgPool,
     nats: &NatsClient,
     job_processor: Box<dyn JobQueueProcessor + Send + Sync>,
@@ -270,7 +270,7 @@ pub async fn migrate_builtins(
     ctx.update_tenancy(Tenancy::new(*workspace.pk()));
     ctx.blocking_commit().await?;
 
-    builtins::migrate(&ctx, selected_test_builtin_schemas).await?;
+    builtins::migrate_local(&ctx, selected_test_builtin_schemas).await?;
 
     ctx.blocking_commit().await?;
